+++ inst/runit_tests/runit-is_files_current.R
@@ -11,15 +11,18 @@ test_is_files_current <- function() {
     RUnit::checkTrue(is_files_current(p3, newer_than = 1, units = "days",
                                       within = 4, within_units = "secs"))
     # more files
-    RUnit::checkTrue(is_files_current(p1, p2, p3, newer_than = 1, units = "days",
+    RUnit::checkTrue(is_files_current(p1, p2, p3, newer_than = 1,
+                                      units = "days",
                      within = 4, within_units = "secs"))
     # within not TRUE
-    RUnit::checkTrue(!is_files_current(p1, p2, p3, newer_than = 1, units = "days",
+    RUnit::checkTrue(!is_files_current(p1, p2, p3, newer_than = 1,
+                                       units = "days",
                                        within = 1, within_units = "secs"))
     # newer not TRUE
-    RUnit::checkTrue(!is_files_current(p1, p2, p3, newer_than = 1, units = "secs",
+    RUnit::checkTrue(!is_files_current(p1, p2, p3, newer_than = 1,
+                                       units = "secs",
                                        within = 4, within_units = "secs"))
 }
 if (interactive()) {
     test_is_files_current()
-} 
+}